refactor and pause menu template

+++ Kursovaya/MusicSource.cs
@@ -12,7 +12,7 @@ public partial class MusicSource : AudioStreamPlayer
 
     public override void _Process(double delta)
     {
-		CurrentTime = (double)((GetPlaybackPosition() + AudioServer.GetTimeSinceLastMix() - AudioServer.GetOutputLatency()) / 60f * BPM);
+		CurrentTime = (GetPlaybackPosition() + AudioServer.GetTimeSinceLastMix() - AudioServer.GetOutputLatency()) / 60f * BPM;
         GetTree().CallGroup("paths", "UpdateTime", CurrentTime);
     }
 }
